Add:(Moderator Controller Added)

+++ src/main/java/com/collegeBook/CollegeBook/service/impl/UserServiceImpl.java
@@ -1,7 +1,9 @@
 package com.collegeBook.CollegeBook.service.impl;
 
+import com.collegeBook.CollegeBook.constant.StringConstant;
 import com.collegeBook.CollegeBook.entity.Role;
 import com.collegeBook.CollegeBook.entity.User;
+import com.collegeBook.CollegeBook.enums.RoleEnum;
 import com.collegeBook.CollegeBook.exception.AppException;
 import com.collegeBook.CollegeBook.pojo.auth.SignInRequest;
 import com.collegeBook.CollegeBook.pojo.auth.SignUpRequest;
@@ -13,8 +15,7 @@ import com.collegeBook.CollegeBook.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -49,7 +50,7 @@ public class UserServiceImpl implements UserService {
         if(user.getPassword().equals(signInRequest.getPassword())){
             return "Login SuccessFul";
         }else{
-            return "Login Failed";
+            throw new AppException("Incorrect Username or password");
         }
 
     }
@@ -60,7 +61,7 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByUserName(username).orElseThrow(()->new AppException("User not found"));
 
         List<String> roles = user.getRoles().stream().map(role->role.getName()).collect(Collectors.toList());
-        return new UserResponse(user.getFirstName(),user.getLastName(),user.getUserName(),roles);
+        return new UserResponse(user.getId(),user.getFirstName(),user.getLastName(),user.getUserName(),roles);
     }
 
     @Override
@@ -83,10 +84,71 @@ public class UserServiceImpl implements UserService {
     @Override
     public String deleteUser(String username) {
 
+        Boolean a =userRepository.findUserWithRoleAdminOrModerator(username);
+
+        if(userRepository.findUserWithRoleAdminOrModerator(username)){
+            throw new AppException("Admin or Moderator cannot be deleted");
+        }
+
+
         User user= userRepository.findByUserName(username).orElseThrow(()->new AppException("user not found"));
         userRepository.delete(user);
         return "User with Id "+user.getId()+" deleted Successfully";
     }
+
+    @Override
+    public List<UserResponse> getUsers() {
+       List<User> users = userRepository.findAll();
+       List<UserResponse> userResponses = new ArrayList<>();
+       for(User user : users){
+
+           List<String> roles = user.getRoles().stream().map(role -> role.getName()).collect(Collectors.toList());
+
+           userResponses.add(new UserResponse(user.getId(),user.getFirstName(),user.getLastName(),user.getUserName(),roles));
+       }
+
+        return userResponses;
+    }
+
+    @Override
+    public String makeModerator(String username) {
+
+        User user = userRepository.findByUserName(username).orElseThrow(()->new AppException("User not found"));
+
+        Role role = roleRepository.findByName(RoleEnum.MODERATOR.name()).orElseThrow(()->new AppException("Role not found"));
+        user.getRoles().add(role);
+        userRepository.save(user);
+        return StringConstant.MADE_MODERATOR;
+    }
+
+    @Override
+    public String removeModerator(String username) {
+
+        User user = userRepository.findByUserName(username).orElseThrow(()->new AppException("User not found"));
+
+        Role role = roleRepository.findByName(RoleEnum.MODERATOR.name()).orElseThrow(()->new AppException("Role not found"));
+
+        user.getRoles().remove(role);
+        userRepository.save(user);
+        return StringConstant.REMOVE_MODERATOR;
+    }
+
+    @Override
+    public List<UserResponse> getModerators(String roleName) {
+
+        List<User> moderators = userRepository.getModerators(RoleEnum.MODERATOR.name());
+
+        List<UserResponse> userResponses = new ArrayList<>();
+
+        for(User user:moderators){
+
+            List<String> roles = user.getRoles().stream().map(role -> role.getName()).collect(Collectors.toList());
+
+            userResponses.add(new UserResponse(user.getId(),user.getFirstName(),user.getLastName(),user.getUserName(),roles));
+        }
+        return userResponses;
+    }
+
 
     public void validateUserName(String username) throws AppException {
